refacto: scroll within table in dataset explorer
